First publish!

+++ deploy.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-# I know set -e exists, but I like this better
-
 function usage() {
     cat <<EOF >&2
 Usage: ./deploy.sh [flags]
@@ -10,18 +8,23 @@ Usage: ./deploy.sh [flags]
 
 Flags:
   -h    Display this message and quit
+  -l    Only lint and build angular-pane-manager
   -n    Perform a dry-run, building but skipping the deploy step
   -u    Same as -h
 EOF
 }
 
 noop=''
+lib=''
 
-while getopts "hnu" opt; do
+while getopts "hlnu" opt; do
     case $opt in
         h|u)
             usage
             exit 0
+            ;;
+        l)
+            lib='-l'
             ;;
         n)
             noop=1
@@ -40,9 +43,9 @@ if (( $# != 0 )); then
     exit 1
 fi
 
-[ -f .deploy.env ] && . ./.deploy.env
+[[ -f .deploy.env ]] && . ./.deploy.env
 
-if [ -z "$ADDRESS" ]; then
+if [[ -z "$ADDRESS" ]]; then
     echo $'\x1b[1;38;5;1mNo $ADDRESS found; aborting.\x1b[m'
     exit -1
 fi
@@ -50,13 +53,24 @@ fi
 set -e
 
 echo $'\x1b[1mLinting...\x1b[m'
-./lint.sh
+./lint.sh $lib
 
 echo $'\x1b[1mBuilding...\x1b[m'
-ng build --prod
-mv dist/angular-pane-manager-example/index{.prod,}.html
+
+if [[ -n "$lib" ]]; then
+    ng build --prod angular-pane-manager
+else
+    ng build --prod
+    mv dist/angular-pane-manager-example/index{.prod,}.html
+fi
 
 if [[ -z "$noop" ]]; then
     echo $'\x1b[1mDeploying...\x1b[m'
-    rsync -rcp --progress --delete dist/angular-pane-manager-example/ "$ADDRESS"
+
+    if [[ -n "$lib" ]]; then
+        cd dist/angular-pane-manager
+        yarn publish --access=public
+    else
+        rsync -rcp --progress --delete dist/angular-pane-manager-example/ "$ADDRESS"
+    fi
 fi
